Added some basic framework
Created the core interfaces for tooltip components as well as the value and conditionals associated with them. Also set up a basic configuration file so users can do mod-level configurations without needing to touch the more modpack-developer focused settings. An example setting for whether or not to use a custom font renderer is in place (inactive)